mt_server/requirements.txt
amqp==2.2.1
Babel==2.5.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
cloudinary==1.8.0
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-health-check==2.4.0
django-phonenumber-field==1.3.0
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.1.2
elasticsearch==5.4.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
mock==2.0.0
oauthlib==2.0.4
openapi-codec==1.3.2
pbr==3.1.1
phonenumberslite==8.8.1
psycopg2==2.7.3.1
PyJWT==1.5.3
PySocks==1.6.7
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
websocket-client==0.44.0
